+++ examples/hyperbolic_serre_green_naghdi_1d/hyperbolic_serre_green_naghdi_soliton_reflecting_relaxation.jl
@@ -1,5 +1,6 @@
 using OrdinaryDiffEqLowStorageRK
 using DispersiveShallowWater
+using SummationByPartsOperators: MattssonNordström2004, derivative_operator
 
 ###############################################################################
 # Semidiscretization of the hyperbolic Serre-Green-Naghdi equations
@@ -9,7 +10,7 @@ equations = HyperbolicSerreGreenNaghdiEquations1D(bathymetry_type = bathymetry_f
                                                   gravity = 9.81)
 
 initial_condition = initial_condition_soliton
-boundary_conditions = boundary_condition_periodic
+boundary_conditions = boundary_condition_reflecting
 
 # create homogeneous mesh
 coordinates_min = -50.0
@@ -17,9 +18,13 @@ coordinates_max = 50.0
 N = 512
 mesh = Mesh1D(coordinates_min, coordinates_max, N)
 
-# create solver with periodic SBP operators of accuracy order 4
-accuracy_order = 4
-solver = Solver(mesh, accuracy_order)
+# create solver with SBP operators of accuracy order 2
+accuracy_order = 2
+D1 = derivative_operator(MattssonNordström2004();
+                         derivative_order = 1, accuracy_order,
+                         xmin = xmin(mesh), xmax = xmax(mesh), N = N)
+
+solver = Solver(D1)
 
 # semidiscretization holds all the necessary data structures for the spatial discretization
 semi = Semidiscretization(mesh, equations, initial_condition, solver,
@@ -34,8 +39,11 @@ summary_callback = SummaryCallback()
 analysis_callback = AnalysisCallback(semi; interval = 100,
                                      extra_analysis_errors = (:conservation_error,),
                                      extra_analysis_integrals = (waterheight_total,
+                                                                 energy_total_modified,
                                                                  entropy_modified))
-callbacks = CallbackSet(analysis_callback, summary_callback)
+relaxation_callback = RelaxationCallback(invariant = entropy_modified)
+# Always put relaxation_callback before analysis_callback to guarantee conservation of the invariant
+callbacks = CallbackSet(relaxation_callback, analysis_callback, summary_callback)
 
 saveat = range(tspan..., length = 100)
 # optimized time integration methods like this one are much more efficient