+++ src/equations/serre_green_naghdi_1d.jl
@@ -275,7 +275,7 @@ function source_terms_manufactured(q, x, t,
 end
 
 """
-    initial_condition_manufactured_reflecting(x, t, equations::SerreGreenNaghdiEquations1D{BathymetryFlat}, mesh)
+    initial_condition_manufactured_reflecting(x, t, equations::SerreGreenNaghdiEquations1D, mesh)
 
 A smooth manufactured solution for reflecting boundary conditions in combination
 with [`source_terms_manufactured_reflecting`](@ref).